fix build issue
