+++ app/routes/nachlass.erbschein.erbfolge.ergebnis.$.tsx
@@ -1,6 +1,6 @@
 import type { LoaderFunctionArgs } from "react-router";
 import { erbfolgeResultExtras } from "~/domains/nachlass/erbschein/erbfolge/resultExtras";
-import { loadResultData } from "~/routes/shared/newEngineResult";
+import { loadResultData } from "~/routes/shared/newEngineResult.server";
 
 export { ResultPage as default } from "~/routes/shared/components/ResultPage";
